+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/text/AnyPrettyNode.java
@@ -5,10 +5,12 @@ import com.oracle.truffle.api.dsl.Fallback;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.interop.InteropLibrary;
 import com.oracle.truffle.api.interop.UnsupportedMessageException;
+import com.oracle.truffle.api.library.CachedLibrary;
 import com.oracle.truffle.api.nodes.Node;
 import org.enso.interpreter.dsl.BuiltinMethod;
 import org.enso.interpreter.runtime.callable.atom.Atom;
 import org.enso.interpreter.runtime.callable.atom.AtomConstructor;
+import org.enso.interpreter.runtime.callable.atom.StructsLibrary;
 import org.enso.interpreter.runtime.data.text.Text;
 
 @BuiltinMethod(type = "Any", name = "pretty", description = "Generic pretty printing function.")
@@ -26,11 +28,12 @@ public abstract class AnyPrettyNode extends Node {
   public abstract Text execute(Object self);
 
   @Specialization
-  Text doAtom(Atom at) {
-    if (at.getFields().length == 0) {
+  Text doAtom(Atom at, @CachedLibrary(limit = "10") StructsLibrary structs) {
+    var fields = structs.getFields(at);
+    if (fields.length == 0) {
       return consName(at.getConstructor());
     } else {
-      return doComplexAtom(at);
+      return doComplexAtom(at, fields);
     }
   }
 
@@ -50,20 +53,20 @@ public abstract class AnyPrettyNode extends Node {
   }
 
   @CompilerDirectives.TruffleBoundary
-  private Text doComplexAtom(Atom atom) {
+  private Text doComplexAtom(Atom atom, Object[] fields) {
     Text res = Text.create("(", consName(atom.getConstructor()));
     res = Text.create(res, " ");
     try {
-      res = Text.create(res, showObject(atom.getFields()[0]));
+      res = Text.create(res, showObject(fields[0]));
     } catch (UnsupportedMessageException e) {
-      res = Text.create(res, atom.getFields()[0].toString());
+      res = Text.create(res, fields[0].toString());
     }
-    for (int i = 1; i < atom.getFields().length; i++) {
+    for (int i = 1; i < fields.length; i++) {
       res = Text.create(res, " ");
       try {
-        res = Text.create(res, showObject(atom.getFields()[i]));
+        res = Text.create(res, showObject(fields[i]));
       } catch (UnsupportedMessageException e) {
-        res = Text.create(res, atom.getFields()[i].toString());
+        res = Text.create(res, fields[i].toString());
       }
     }
     res = Text.create(res, ")");